Move length check func to util.py
