+++ unit-tests/test_util.py
@@ -13,7 +13,7 @@ sys.path.append("..")
 from util import get_validated_static_path
 from util import is_email_in_domains
 from util import unicode_secure_filename
-
+from util import length_check
 
 class UtilTest(TestCase):
     def test_is_email_in_domains(self):
@@ -187,3 +187,18 @@ class UtilTest(TestCase):
                 "uu",
             )
         )
+
+    def test_length_check_empty(self):
+        """Test that an empty banner message is identified as invalid."""
+        _, is_valid = length_check("")
+        self.assertFalse(is_valid)
+
+    def test_length_check_too_long(self):
+        """Test that a too-long banner message is identified as invalid."""
+        _, is_valid = length_check("a" * 257)  # 257 characters long
+        self.assertFalse(is_valid)
+
+    def test_length_check_valid(self):
+        """Test that a valid banner message is accepted."""
+        _, is_valid = length_check("Maintenance scheduled from 1 July 7:00 to 2 July 9:00.")
+        self.assertTrue(is_valid)
