+++ unit-tests/unit_test.py
@@ -5,12 +5,10 @@ __license__   = 'GPLv3, see LICENSE'
 
 from unittest import makeSuite, TestSuite
 
-from test_admin import AdminTest
 from test_util import UtilTest
 
 
 def suite():
     test_suite = TestSuite()
-    test_suite.addTest(makeSuite(AdminTest))
     test_suite.addTest(makeSuite(UtilTest))
     return test_suite
